Fix for #172
The problem was that the binding of the filterablebeforefilter event did
not get renewed when changing to another category, although all other
events do rebind fine. Changed to listening for the input event of the
text box instead, which requires an explicit timeout (which otherwise is
implemented by the filterable construct).